pyproject.toml
[tool.poetry]
name = "chatbot-task"
version = "0.1.0"
description = ""
authors = ["Ali Kemal Yıldırım <akemalyildirim1@gmail.com>"]
readme = "README.md"
packages = [
    { include = "src/" },
    {include = "*", from = "src"},
]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = {extras = ["all"], version = "^0.111.0"}
alembic = "^1.13.1"
asyncpg = "^0.29.0"
pydantic-settings = "^2.3.3"
pydantic = "^2.7.4"
sqlalchemy = {extras = ["mypy"], version = "^2.0.30"}
aiohttp = "^3.9.5"
pypdf2 = "^3.0.1"
weaviate-client = "^4.6.5"
tokenizers = "^0.19.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.2"
pytest-mock = "^3.14.0"
pre-commit = "^3.7.1"
ruff = "^0.4.10"
mypy = "^1.10.0"
coverage = "^7.5.3"
pytest-asyncio = "^0.23.7"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.ruff]
line-length = 88
exclude = ["alembic", "tests"]

[tool.ruff.lint]
select = [
    # pydocstyle
    "D",
    # pycodestyle
    "E",
    # Pyflakes
    "F",
    # pyupgrade
    "UP",
    # flake8-bugbear
    "B",
    # flake8-simplify
    "SIM",
    # isort
    "I",
]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
docstring-code-format = true

[tool.ruff.lint.pydocstyle]
convention = "google"

[tool.mypy]
plugins = ["pydantic.mypy", "sqlalchemy.ext.mypy.plugin"]
ignore_missing_imports = true
strict_optional = false
exclude = ["alembic", ]

[tool.pytest.ini_options]
asyncio_mode = "auto"

[tool.coverage.run]
branch = true
source = ["src"]


[tool.coverage.report]
skip_empty = true
show_missing = true
fail_under = 100
